Moves angle/abs test to test_torch
